+++ src/neuron.cc
@@ -6,11 +6,11 @@ Neuron::Neuron(unsigned int input_size, unsigned int activation_function): input
     this->weighted_sum_ = 0.0f;
     this->nodeVal_ = 0.0f;
     
-    this->bias_ = ((float)rand() / (float)RAND_MAX) * 2 - 1;
+    this->bias_ = ((double)rand() / (double)RAND_MAX) * 2 - 1;
     this->weights_.reserve(this->input_size_);
 
     for(int i = 0; i < this->input_size_; i++){
-        this->weights_.push_back(((float)rand() / (float)RAND_MAX) * 2 - 1);
+        this->weights_.push_back(((double)rand() / (double)RAND_MAX) * 2 - 1);
     }
 }
 
@@ -18,15 +18,15 @@ unsigned int Neuron::get_input_size(){
     return this->input_size_;
 }
 
-float Neuron::get_output(){
+double Neuron::get_output(){
     return this->output_;
 }
 
-float Neuron::get_weighted_sum(){
+double Neuron::get_weighted_sum(){
     return this->weighted_sum_;
 }
 
-float Neuron::get_weight_value(unsigned int weight){
+double Neuron::get_weight_value(unsigned int weight){
     return this->weights_[weight];
 }
 
@@ -34,13 +34,13 @@ void Neuron::reset_nodeVal(){
     this->nodeVal_ = 0.0f;
 }
 
-float Neuron::propogate(const std::vector<float>& inputs){
+double Neuron::propogate(const std::vector<double>& inputs){
     if(inputs.size() != this->input_size_){
         std::cerr << "Input size does not match neuron input size" << std::endl;
         return 0.0f;
     }
 
-    float sum = this->bias_;
+    double sum = this->bias_;
 
     for (int i = 0; i < this->input_size_; i++){
         sum += inputs[i] * this->weights_[i];
@@ -50,20 +50,20 @@ float Neuron::propogate(const std::vector<float>& inputs){
     return this->output_;
 }
 
-void Neuron::update_nodeVal(float nodeVal){
+void Neuron::update_nodeVal(double nodeVal){
     nodeVal_ += nodeVal;
 }
 
-void Neuron::update_weights(unsigned int weight, float input, float learning_rate){
+void Neuron::update_weights(unsigned int weight, double input, double learning_rate){
     this->weights_[weight] -= learning_rate * nodeVal_ * input;
 }
 
-void Neuron::update_bias(float learning_rate){
+void Neuron::update_bias(double learning_rate){
     bias_ -= learning_rate * nodeVal_;
 }
 
 
-float Neuron::activation_function(float input){
+double Neuron::activation_function(double input){
     if(activation_function_ == 0){
         //ReLU
         if(input < 0.0f){
@@ -78,7 +78,7 @@ float Neuron::activation_function(float input){
 }
 
 
-float Neuron::activation_function_derivative(float input){
+double Neuron::activation_function_derivative(double input){
     if(activation_function_ == 0){
         //ReLU
         if(input < 0.0f){
@@ -88,7 +88,7 @@ float Neuron::activation_function_derivative(float input){
     }
     else {
         //Sigmoid
-        float activation = activation_function(input);
+        double activation = activation_function(input);
         return activation * (1.0f - activation);
     }
 }
